requirements.txt
pandas>=1.3
scikit-learn>=1.0
pyyaml>=6.0
joblib>=1.1
mlflow>=2.0
dvc>=2.0
